Change Speed to work in microseconds.

+++ source/frontends/common2/speed.cpp
@@ -19,13 +19,13 @@ namespace common2
     myStartCycles = g_nCumulativeCycles;
   }
 
-  uint64_t Speed::getCyclesAtFixedSpeed(const size_t microseconds) const
+  uint64_t Speed::getCyclesAtFixedSpeed(const uint64_t microseconds) const
   {
     const uint64_t cycles = static_cast<uint64_t>(microseconds * g_fCurrentCLK6502 * 1.0e-6) + g_nCpuCyclesFeedback;
     return cycles;
   }
 
-  uint64_t Speed::getCyclesTillNext(const size_t microseconds)
+  uint64_t Speed::getCyclesTillNext(const uint64_t microseconds)
   {
     if (myFixedSpeed || g_bFullSpeed)
     {
